backend/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-cors==6.0.1
flask-jwt-extended==4.7.1
python-dotenv==1.2.1
flask-swagger-ui==5.21.0
SQLAlchemy==2.0.44
PyJWT==2.10.1